requirements.txt
flask
requests
python-dotenv
langchain-core
langchain-google-genai
langchain-agents
